Move ops classes/files into the data directory to be more close to the actual data objects
