+++ map_engraver/data/geo/geo_coordinate_transformers.py
@@ -2,7 +2,7 @@ from typing import List
 
 from pyproj import proj, Transformer
 
-from map_engraver.transformers.geo_coordinate import GeoCoordinate
+from map_engraver.data.geo.geo_coordinate import GeoCoordinate
 
 
 def transform_geo_coordinate_to_new_crs(
