+++ map_engraver/data/geo_canvas_ops/geo_canvas_transformers.py
@@ -2,11 +2,11 @@ import pyproj
 from typing import Tuple, Callable, Optional
 
 from map_engraver.canvas.canvas_coordinate import CanvasCoordinate
-from map_engraver.transformers.geo_coordinate import GeoCoordinate
-from map_engraver.transformers.geo_canvas_scale import GeoCanvasScale
+from map_engraver.data.geo.geo_coordinate import GeoCoordinate
+from map_engraver.data.geo_canvas_ops.geo_canvas_scale import GeoCanvasScale
 
 
-def build_geo_to_canvas_transformer(
+def build_transformer(
         # The Coordinate Reference System for plotting coordinate data onto
         # the canvas.
         crs: pyproj.CRS,
