+++ tests/data/osm_shapely/data.osm
