+++ tests/data/osm_shapely/incomplete_data.osm
